[zubair]  remaining completed fully

+++ backend/src/controllers/notes.ts
@@ -2,11 +2,15 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
   try {
-    const notes = await NoteModel.find().exec();
-    res.status(200).json(notes);
+      assertIsDefined(authenticatedUserId);
+
+      const notes = await NoteModel.find({userId: authenticatedUserId}).exec();
+      res.status(200).json(notes);
   } catch (error) {
     next(error);
   }
@@ -14,7 +18,11 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
+
     try {
+        assertIsDefined(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400, "Invalid note Id")
         }
@@ -22,6 +30,10 @@ export const getNote: RequestHandler = async (req, res, next) => {
 
         if (!note) {
             throw createHttpError(404, "Note not found");
+        }
+
+        if (!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401, "You cannot access this note")
         }
         res.status(200).json(note);
         
@@ -38,13 +50,17 @@ interface CreateNoteBody {
 export const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
-    // const {title, text} = req.body;
+  const authenticatedUserId = req.session.userId;
+
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!title) {
         throw createHttpError(400, "not must have title")  ;
     }
     const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
       title: title,
       text: text,
     });
@@ -67,7 +83,11 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     const noteId =  req.params.noteId;
     const newTitle = req.body.title;
     const newText = req.body.text;
+    const authenticatedUserId = req.session.userId;
+
     try {
+        assertIsDefined(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400, "Invalid note Id")
         }
@@ -80,6 +100,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
         if (!note) {
             throw createHttpError(404, "Note not found");
+        }
+
+        if (!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401, "You cannot access this note")
         }
 
         note.title = newTitle;
@@ -96,8 +120,12 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
 export const deleteNote: RequestHandler = async(req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
+
 
     try {
+        assertIsDefined(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id");
         }
@@ -105,6 +133,10 @@ export const deleteNote: RequestHandler = async(req, res, next) => {
 
         if (!note) {
             throw createHttpError(404, "Note not found");
+        }
+
+        if (!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401, "You cannot access this note")
         }
 
         await note.deleteOne();
